Arrange app and tasks so they are separately accessible
Fix flake8 errors
