+++ tests/flake8.cfg
@@ -43,6 +43,7 @@ quotes: '
 #     D400 - First line should end with a period.
 #     D401 - First line should be in imperative mood.
 #     D402 - First line should not be the function's "signature".
+#     D403 - First word of the first line should be properly capitalized
 #
 #   Identifier errors
 #   ~~~~~~~~~~~~~~~~~
@@ -53,4 +54,4 @@ quotes: '
 #     N803 - Argument name should be lowercase.
 #     N806 - Variable in function should be lowercase.
 #     N812 - Lowercase imported as non lowercase.
-ignore: D100,D101,D102,D103,D104,D105,D200,D201,D202,D203,D204,D205,D400,D401,D402,E123,E226,E241,N802,N803,N806,N812
+ignore: D100,D101,D102,D103,D104,D105,D200,D201,D202,D203,D204,D205,D400,D401,D402,D403,E123,E226,E241,N802,N803,N806,N812
